Connected With H2 DB

+++ src/main/java/com/CRUD/myproject/controller/UserController.java
@@ -44,7 +44,7 @@ public class UserController {
     
     @DeleteMapping("/{id}")
     public String deleteUser(@PathVariable int id){
-        boolean deleteResult =  service.deleteCurrentUser(id);
+        boolean deleteResult =  service.deleteCurrentUser((long) id);
         return deleteResult ? "true" : "false";
     }
 }
